requirements.txt
# Kafka
confluent-kafka==2.3.0

# Elasticsearch
elasticsearch==8.13.0

# Redis
redis==5.0.3

# S3 / MinIO
minio==7.2.2

# ML & Anomaly Detection
scikit-learn==1.4.2
xgboost==2.0.3
torch==2.2.2
numpy==1.26.4
pandas==2.2.1
shap==0.45.0
imbalanced-learn==0.12.0

# Agent Framework
langchain==0.1.16
langgraph==0.0.40

# NLP for log parsing
transformers==4.39.3

# HTTP & Async
httpx==0.27.0
aiohttp==3.9.5
fastapi==0.110.1
uvicorn==0.29.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.0
loguru==0.7.2
schedule==1.2.1
colorama==0.4.6

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6

